feat(CLN-002): split app.ts and consolidation-loop.ts into focused modules
- Extract TYPE_DIRS, ensureDataDirectories, resolveFilePath to lib/file-utils.ts
- Create routes/system.ts, ingestion.ts, memory.ts, consolidation.ts, operations.ts
- app.ts reduced from ~449 to 88 lines (app creation + middleware + route mounting)
- Extract loadSeedFromDisk, loadClusterMemberFiles, enrichCandidatesWithFiles, getExistingInsights to consolidation-loaders.ts
- Extract runConsolidationCycle, runConsolidationDryRun to consolidation-cycle.ts
- consolidation-loop.ts reduced to 206 lines (loop orchestration, reconcileOnStartup, buildConsolidationDeps)
- All existing exports re-exported to preserve test and consumer compatibility
- 842 tests pass, typecheck clean

Closes #99
